moved ElasticDBTask back to elastic_tasks
